Added a bunch of tests and restructured the package layout
