+++ Chess/src/test/controller/AboutControllerTest.java
@@ -1,0 +1,5 @@
+package controller;
+
+public class AboutControllerTest {
+
+}
