+++ Chess/src/test/controller/LoginControllerTest.java
@@ -1,0 +1,5 @@
+package controller;
+
+public class LoginControllerTest {
+
+}
