--- Chess/src/test/core/RunnerTest.java
@@ -1,13 +1,0 @@
-package core;
-
-import core.client.Runner;
-
-/**
- * Base test class for {@link Runner}
- * @author Brisbin
- *
- */
-public class RunnerTest 
-{
-
-}
