--- Chess/src/test/core/ServerTest.java
@@ -1,5 +1,0 @@
-package core;
-
-public class ServerTest {
-
-}
