+++ Chess/src/test/core/client/ClientTest.java
@@ -1,4 +1,4 @@
-package core;
+package core.client;
 
 public class ClientTest {
