+++ Chess/src/test/core/client/ConnectionTest.java
@@ -1,0 +1,5 @@
+package core.client;
+
+public class ConnectionTest {
+
+}
