+++ Chess/src/test/core/client/MessageTest.java
@@ -1,0 +1,5 @@
+package core.client;
+
+public class MessageTest {
+
+}
