+++ Chess/src/test/core/client/ModelTest.java
@@ -1,4 +1,4 @@
-package core;
+package core.client;
 
 public class ModelTest {
