+++ Chess/src/test/core/client/ProductInfoTest.java
@@ -1,0 +1,5 @@
+package core.client;
+
+public class ProductInfoTest {
+
+}
