+++ Chess/src/test/core/client/game/BishopTest.java
@@ -7,6 +7,7 @@ import java.awt.Color;
 import org.junit.Test;
 
 import core.client.Coordinate;
+import core.client.game.Bishop;
 
 /**
  * Base test class for {@link Bishop}
