+++ Chess/src/test/core/client/game/EmptyPieceTypeTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
+import core.client.game.EmptyPiece;
+
 public class EmptyPieceTypeTest 
 {
 	@Test
