+++ Chess/src/test/core/client/game/KingTest.java
@@ -7,6 +7,7 @@ import java.awt.Color;
 import org.junit.Test;
 
 import core.client.Coordinate;
+import core.client.game.King;
 
 /**
  * Base test class for {@link King}
